Change Blur arg from double -> float

+++ src/Carbon.Media/Processors/MediaTransformation.cs
@@ -196,7 +196,7 @@ namespace Carbon.Media
             return this;
         }
 
-        public MediaTransformation Blur(double radius)
+        public MediaTransformation Blur(float radius)
         {
             Apply(new BlurFilter(radius));
